free val->result after failing parse
Make behaviour consistent with failing execution. Although I don't see
how a result could be left over, given that the string could not have
changed (pathcomp_set() frees the value and sets it again) _and_ it
could almost certainly not fail to parse second time round, it is better
not to rely on any such assumption. Also, it's not wrong to free the
result: it cannot be relied upon anyway.