do not check existence of lib/ when running test_mkdir.c
This check was included to avoid polluting other directories than test/
with a directory lib/. (test_mkdir() does not remove test/lib/ after
testing because test/lib/ contains data files that must not be erased.)
But in VPATH builds, test/lib/ does not exist in the build dir, so we
cannot assume its existence. A better solution would probably be to
perform directory creation tests in a temporary directory.