pass value_t pointer into att_new() and related functions
This is a change in preparation of another commit, which will introduce
integer values. With this change, att_new() and related functions no
longer allocate a value object themselves; instead, they rely on the
caller passing in a pointer to a value_t object. This allows the caller
to decide which kind of value they want.

A side effect of this change is that values are now constructed in
pathcomp.c, but destroyed in att.c (in att_free(), at the time the
attribute is destroyed). This asymmetry is documented in att.c, and
seems justified to me.